bug-bounty-hunter/scheduler/monitoring.py
"""
Scheduler - Automated continuous monitoring and scanning
"""
from apscheduler.schedulers.asyncio import AsyncIOScheduler
from apscheduler.triggers.interval import IntervalTrigger
from datetime import datetime
from loguru import logger
import asyncio
from typing import List, Dict
import yaml


class BountyHuntingScheduler:
    """
    Manages scheduled tasks for continuous bounty hunting:
    - Platform monitoring for new programs
    - Periodic rescanning of targets
    - Health checks
    - Notification delivery
    """
    
    def __init__(self, config: Dict):
        self.config = config
        self.scheduler = AsyncIOScheduler()
        self.running = False
        
        # Import managers (would be passed in production)
        # self.platform_manager = platform_manager
        # self.scan_manager = scan_manager
        
    async def check_new_programs(self):
        """Check all platforms for new programs"""
        logger.info("Checking for new bug bounty programs...")
        
        try:
            # TODO: Implement with actual platform manager
            # new_programs = await self.platform_manager.check_new_programs()
            
            # For now, simulate
            new_programs = []
            
            if new_programs:
                logger.info(f"Found {len(new_programs)} new programs!")
                # TODO: Notify user
                # TODO: Auto-analyze and queue high-priority targets
            else:
                logger.info("No new programs found")
                
        except Exception as e:
            logger.error(f"Error checking new programs: {e}")
    
    async def rescan_programs(self):
        """Rescan existing programs for updates"""
        logger.info("Rescanning existing programs...")
        
        try:
            # TODO: Implement rescanning logic
            # Get all active programs
            # Check for scope changes
            # Re-run scans if needed
            pass
        except Exception as e:
            logger.error(f"Error rescanning programs: {e}")
    
    async def cleanup_old_data(self):
        """Clean up old scan results and logs"""
        logger.info("Cleaning up old data...")
        
        try:
            # TODO: Implement cleanup
            # Remove old scan results
            # Archive old logs
            # Clean temporary files
            pass
        except Exception as e:
            logger.error(f"Error during cleanup: {e}")
    
    async def check_worker_health(self):
        """Check health of all worker nodes"""
        logger.debug("Checking worker health...")
        
        try:
            # TODO: Implement health check
            # Ping all workers
            # Mark unresponsive workers as offline
            # Notify if workers are down
            pass
        except Exception as e:
            logger.error(f"Error checking worker health: {e}")
    
    async def generate_daily_report(self):
        """Generate daily activity report"""
        logger.info("Generating daily report...")
        
        try:
            # TODO: Implement reporting
            # Collect statistics
            # Generate report
            # Send notifications
            pass
        except Exception as e:
            logger.error(f"Error generating report: {e}")
    
    def start(self):
        """Start the scheduler"""
        if self.running:
            logger.warning("Scheduler is already running")
            return
        
        interval = self.config.get('monitoring', {}).get('check_interval', 3600)
        
        # Schedule tasks
        self.scheduler.add_job(
            self.check_new_programs,
            trigger=IntervalTrigger(seconds=interval),
            id='check_programs',
            name='Check for new programs',
            replace_existing=True
        )
        
        self.scheduler.add_job(
            self.rescan_programs,
            trigger=IntervalTrigger(hours=12),
            id='rescan_programs',
            name='Rescan existing programs',
            replace_existing=True
        )
        
        self.scheduler.add_job(
            self.check_worker_health,
            trigger=IntervalTrigger(minutes=5),
            id='health_check',
            name='Worker health check',
            replace_existing=True
        )
        
        self.scheduler.add_job(
            self.cleanup_old_data,
            trigger=IntervalTrigger(days=1),
            id='cleanup',
            name='Clean old data',
            replace_existing=True
        )
        
        self.scheduler.add_job(
            self.generate_daily_report,
            trigger=IntervalTrigger(days=1),
            id='daily_report',
            name='Daily report',
            replace_existing=True
        )
        
        self.scheduler.start()
        self.running = True
        logger.info(f"Scheduler started with {interval}s check interval")
    
    def stop(self):
        """Stop the scheduler"""
        if not self.running:
            return
        
        self.scheduler.shutdown()
        self.running = False
        logger.info("Scheduler stopped")
    
    def get_jobs(self) -> List[Dict]:
        """Get list of scheduled jobs"""
        jobs = []
        for job in self.scheduler.get_jobs():
            jobs.append({
                'id': job.id,
                'name': job.name,
                'next_run': job.next_run_time.isoformat() if job.next_run_time else None
            })
        return jobs


# Example usage
if __name__ == "__main__":
    import sys
    
    # Load config
    with open("../config/config.yaml", "r") as f:
        config = yaml.safe_load(f)
    
    scheduler = BountyHuntingScheduler(config)
    
    async def main():
        scheduler.start()
        
        print("Scheduler started. Scheduled jobs:")
        for job in scheduler.get_jobs():
            print(f"  - {job['name']}: {job['next_run']}")
        
        print("\nPress Ctrl+C to stop")
        
        try:
            # Keep running
            while True:
                await asyncio.sleep(1)
        except KeyboardInterrupt:
            print("\nStopping scheduler...")
            scheduler.stop()
    
    asyncio.run(main())
