bug-bounty-hunter/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Distributed Task Queue
celery==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.9
motor==3.3.2  # Async MongoDB driver
sqlalchemy==2.0.25

# Web Scraping & HTTP
requests==2.31.0
httpx==0.26.0
beautifulsoup4==4.12.3
selenium==4.17.2
playwright==1.41.0

# AI/ML
torch==2.1.2
transformers==4.37.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# Security & Scanning Tools Integration
python-nmap==0.7.1
dnspython==2.5.0

# Scheduling & Monitoring
apscheduler==3.10.4
prometheus-client==0.19.0

# Configuration & Logging
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# API Clients for Bounty Platforms
gql==3.4.1  # GraphQL client for HackerOne
aiohttp==3.9.1

# Notifications
python-telegram-bot==20.7
discord-webhook==1.3.1

# Utilities
click==8.1.7
rich==13.7.0  # Beautiful terminal output
tenacity==8.2.3  # Retry logic
